Feat(routes): implement WiFi connection handler with validation and distance check
